# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Autoservis API",
        "description": "API for managing user login, logout, info",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/user": {
            "get": {
                "tags": [
                    "User"
                ],
                "summary": "Get authenticated user information",
                "description": "Returns the authenticated user data",
                "operationId": "getUser",
                "responses": {
                    "200": {
                        "description": "Successful response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/User"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Not authorized"
                    }
                }
            }
        },
...
        "/api/vehicles": {
            "get": {
                "tags": [
                    "Vehicle"
                ],
                "summary": "Get user's vehicles",
                "description": "Returns the authenticated user data",
                "operationId": "index",
                "responses": {
                    "200": {
                        "description": "Successful response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/User"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Not authorized"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
...
            "User": {
                "title": "User",
                "description": "User",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "tenantId": {
                        "type": "integer",
                        "example": 1
                    },
                    "lastRoleId": {
                        "type": "integer",
                        "example": 1
                    },
                    "firstName": {
                        "type": "string",
                        "example": "John"
                    },
                    "lastName": {
                        "type": "string",
                        "example": "Doe"
                    },
                    "loginName": {
                        "type": "string",
                        "example": "john"
                    },
                    "phone": {
                        "type": "string",
                        "example": "+420777521456"
                    },
                    "email": {
                        "type": "string",
                        "example": "admin@examle.com"
                    },
                    "emailVerifiedAt": {
                        "type": "data",
                        "example": "2025-01-16T13:40:05.000000Z"
                    },
                    "active": {
                        "type": "integer",
                        "example": 1
                    },
                    "createdAt": {
                        "type": "data",
                        "example": "2025-01-16T13:40:05.000000Z"
                    },
                    "updatedAt": {
                        "type": "data",
                        "example": "2025-01-16T13:40:05.000000Z"
                    }
                },
                "type": "object"
            },
...
        }
    },
...
}